touchAI/projectnew/requirements.txt
django-cors-headers
djangorestframework
python-dateutil
python-dotenv
